requirements.txt
# Core dependencies
# tkinter  # Removed - this is a standard library module, not pip-installable
pyttsx3>=2.90
pygame>=2.5.0
Pillow>=10.0.0
python-dotenv>=1.0.0

# AI and API
google-generativeai>=0.3.0
mistralai>=0.0.7
requests>=2.31.0

# Utilities
pyperclip>=1.8.2

# Audio processing (for sound generation script)
scipy>=1.10.0

# Development tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0

# Add PyInstaller to existing requirements.txt
pyinstaller>=6.3.0

#streamlit>=1.31.0  not needed for now could not deploy because streamlit doesn't support tkinter
